+++ hydromt/io.py
@@ -4,6 +4,7 @@ import io
 import logging
 from os.path import abspath, basename, dirname, isfile, join, splitext
 from pathlib import Path
+from typing import Any, Dict, Literal
 
 import dask
 import geopandas as gpd
@@ -228,13 +229,27 @@ def open_mfraster(
 
 
 def open_mfcsv(
-    fns, driver_kwargs, concat_dim, rename_existing_index_to, variable_axis=1
-):
+    fns: Dict[str | int, str | Path],
+    concat_dim: str,
+    driver_kwargs: Dict[str, Any],
+    variable_axis: Literal[0, 1] = 1,
+) -> xr.Dataset:
     """Open multiple csv files as single Dataset.
 
     Arguments
     ---------
-    TBA
+    fns : Dict[str | int, str | Path],
+        Dictionary containing a id -> filename mapping. Here the ids,
+        should correspond to the values of the `concat_dim` dimension.
+    concat_dim : str,
+        name of the dimention that will be created by concatinating
+        all of the supplied csv files.
+    driver_kwargs : Dict[str, Any],
+        Any additional arguments to be passed to pandas' `read_csv` function.
+    variable_axis : Literal[0, 1] = 1,
+        The axis along which your variables are. so if the csvs have the
+        columns as variable names, you would leave this as 1. If the variables
+        are along the index, set this to 0.
 
     Returns
     -------
@@ -246,31 +261,43 @@ def open_mfcsv(
         raise ValueError(f"there is no axis {variable_axis} avaialbe in 2d csv files")
     # we're gonna use the structure of the first file found to check
     # all others agains
-    first_id, first_fn = fns.popitem()
-    first_df = pd.read_csv(first_fn, **driver_kwargs)
+    csv_kwargs = {"index_col": 0}
+    csv_kwargs.update(**driver_kwargs)
+    first_id, first_fn = next(iter(fns.items()))
+    first_df = pd.read_csv(first_fn, **csv_kwargs)
     if variable_axis == 0:
         first_df = first_df.T
-    first_df.rename_axis(rename_existing_index_to, axis=0, inplace=True)
+
     first_index = first_df.index
+
+    if first_df.index.name is None:
+        csv_index_name = "index"
+    else:
+        csv_index_name = first_df.index.name
+
     first_df[concat_dim] = first_id
-    dfs = [first_df]
+    dfs = []
     for id, fn in fns.items():
-        df = pd.read_csv(fn, **driver_kwargs)
+        df = pd.read_csv(fn, **csv_kwargs)
         if variable_axis == 0:
             df = df.T
+
         df[concat_dim] = id
-        df.rename_axis(rename_existing_index_to, axis=0, inplace=True)
 
         if not df.index.equals(first_index):
-            raise RuntimeError(f"file {fn} has inconsistent index: {df.index}")
+            raise RuntimeError(
+                f"file {fn} has inconsistent index: {df.index}"
+                f"Expected {first_index}"
+            )
 
         dfs.append(df)
 
-    all_dfs_combined = pd.concat(dfs, axis=0).set_index(
-        [concat_dim, rename_existing_index_to]
+    all_dfs_combined = (
+        pd.concat(dfs, axis=0).reset_index().set_index([concat_dim, csv_index_name])
     )
-    ds = xr.Dataset.from_dataframe(all_dfs_combined).drop_vars("Unnamed: 0")
-
+    ds = xr.Dataset.from_dataframe(all_dfs_combined)
+    if "Unnamed: 0" in ds.data_vars:
+        ds = ds.drop_vars("Unnamed: 0")
     return ds