+++ docs/changelog.rst
@@ -16,7 +16,7 @@ Added
 - Add support for reading and writing tabular data in ``parquet`` format. (PR #445)
 - Add support for reading model configs in ``TOML`` format. (PR #444)
 - new ``force-overwrite`` option in ``hydromt update`` CLI to force overwritting updated netcdf files. (PR #460)
--
+- add ``open_mfcsv`` function in ``io`` module for combining multiple CSV files into one dataset. (PR #486)
 
 Changed
 -------